+++ frontend/mobile/src/app/tabs/_layout.tsx
@@ -4,7 +4,6 @@ import { Ionicons } from '@expo/vector-icons';
 import { withLayoutContext } from 'expo-router';
 import { createMaterialTopTabNavigator } from '@react-navigation/material-top-tabs';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
-
 const { Navigator } = createMaterialTopTabNavigator();
 
 export const SwipeTabs = withLayoutContext(Navigator);
@@ -29,6 +28,7 @@ export default function TabsLayout() {
 
         tabBarIndicatorStyle: {
           display: 'none',
+          flex: 1,
         },
       }}
     >
@@ -61,6 +61,17 @@ export default function TabsLayout() {
           ),
         }}
       />
+
+      <SwipeTabs.Screen
+        name="profile"
+        options={{
+          title: 'Profile',
+            tabBarIcon: ({ color }: { color: string }) => (
+              <Ionicons name="person-circle-outline" size={24} color={color} />
+          ),
+        }}
+      />
     </SwipeTabs>
+
   );
 }
