+++ frontend/mobile/src/app/tabs/profile.tsx
@@ -1,0 +1,11 @@
+import ProfileView from '@/ui/views/user/ProfileView';
+import { ScreenLayout } from '@/ui/layouts/ScreenLayout';
+import { AppLayout } from '@/ui/layouts/AppLayout';
+
+export default function RestaurantsPage() {
+  return (
+    <ScreenLayout>
+      <ProfileView />
+    </ScreenLayout>
+  );
+}
